refactor the project, now the construct is
Barcode
└─src
|  └─Dectect.cpp
|  └─testenv.cpp
└─include
|  └─testenv.h
└─test
|  └─datas
|    └─example.jpg
└─cmake
|  └─CMakeHelper.cmake
|  └─OpenCVPath.cmake
└─CMakeLists.txt
