+++ CMakeLists.txt
@@ -1,37 +1,35 @@
-# CMakeList.txt: testCmakeOpencv 的 CMake 项目，在此处包括源代码并定义
-# 项目特定的逻辑。
-#
-cmake_minimum_required (VERSION 3.16.6)
+cmake_minimum_required(VERSION 3.16.6)
+project(Barcode)
 
-project ("BarcodeTest")
-include(OpenCVPath.cmake)
-MESSAGE(${OpenCV_DIR})
+include(cmake/OpenCVPath.cmake) #even opencv itself do not use cmake_module_path
+include(cmake/CMakeHelper.cmake)
+
+set(OPEN_DEBUG ON)
+if (OPEN_DEBUG)
+    set(CMAKE_BUILD_TYPE "Debug")
+    #set it to Debug, it will auto choose *d.dll
+else ()
+    set(CMAKE_BUILD_TYPE "Release")
+    #set it to Release, it will auto choose *.dll
+endif ()
+
+add_compile_options("$<$<C_COMPILER_ID:MSVC>:/utf-8>")
+add_compile_options("$<$<CXX_COMPILER_ID:MSVC>:/utf-8>")
 
 set(OpenCV_LIBS "${OpenCV_DIR}/x64/vc16/lib")
 set(OpenCV_BINS "${OpenCV_DIR}/x64/vc16/bin")
-MESSAGE("OpenCV_DIR: " ${OpenCV_DIR})
-MESSAGE("OpenCV_INCLUDE_DIRS: " ${OpenCV_INCLUDE_DIRS})
-MESSAGE("OpenCV_LIBS: " ${OpenCV_LIBS})
-MESSAGE("OpenCV_BINS: " ${OpenCV_BINS})
 
+Output_Paths()
 find_package(OpenCV REQUIRED)
-MESSAGE("OpenCV_DIR: " ${OpenCV_DIR})
-MESSAGE("OpenCV_INCLUDE_DIRS: " ${OpenCV_INCLUDE_DIRS})
-MESSAGE("OpenCV_LIBS: " ${OpenCV_LIBS})
+Output_Paths()
+
+include_directories(include)
 include_directories(${OpenCV_INCLUDE_DIRS})
 link_directories(${OpenCV_LIBS})
-# 将源代码添加到此项目的可执行文件。
-add_executable (${PROJECT_NAME}_Detect "Detect.cpp")
 
-#SET(ENV{CMAKE_INCLUDE_PATH} ${OpenCV_BINS})
-#SET(ENV{CMAKE_LIBRARY_PATH} ${OpenCV_BINS})
-#SET(ENV{PATH} "${OpenCV_BINS}")
-#MESSAGE($ENV{PATH} " System path")
-#SET(ENV{PATH} "$ENV{PATH};${OpenCV_BINS}")
-#MESSAGE($ENV{PATH} " System path")
+add_executable(${PROJECT_NAME}_Detect src/Detect.cpp)
+target_link_libraries(${PROJECT_NAME}_Detect ${OpenCV_LIBS})
 
-#execute_process(
-#  COMMAND ${CMAKE_COMMAND} -E environment
-#)
-# TODO: 如有需要，请添加测试并安装目标。
-target_link_libraries(${PROJECT_NAME}_Detect opencv_world450d)
+add_executable(${PROJECT_NAME}_test src/testenv.cpp)
+target_link_libraries(${PROJECT_NAME}_test ${OpenCV_LIBS})
+
