+++ README.md
@@ -17,13 +17,14 @@
 如果能正常运行example程序就没啥问题了
 
 ### How To use Cmake
-在目录下面建一个`OpenCVPath.cmake`,
+在.CMake下面建一个`OpenCVPath.cmake`,
 里面放
 
 ``` cmake
 set(OpenCV_DIR ${YOUR_OPENCV_DIR})
 # 比如"C:/repo/opencv_build/Install/opencv"
 ```
+
 同时确认环境变量配置了.
 然后用visual studio-文件-打开-cmake,打开`CMakeLists.txt`.
 即可跑通`BarCodeTest_Dected.exe`
