+++ cmake/CMakeHelper.cmake
@@ -1,0 +1,9 @@
+function(Output_variable arg)
+    MESSAGE("${arg}: " ${${arg}})
+endfunction()
+function(Output_Paths)
+    Output_variable(OpenCV_DIR)
+    Output_variable(OpenCV_INCLUDE_DIRS)
+    Output_variable(OpenCV_LIBS)
+    Output_variable(OpenCV_BINS)
+endfunction()
