+++ src/testenv.cpp
@@ -1,6 +1,5 @@
 #include "opencv2/highgui/highgui.hpp"
 #include "testenv.h"
-#include <string>
 using namespace cv;
 
 int main(int argc, char** argv) {
